requirements.txt
# KeyChaser - Malware C2 Sinkhole Framework
# Core dependencies for the sinkhole and dashboard

# Web Framework & ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6

# Database & ORM
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.0

# Data Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Cryptography (for malware payload decryption)
pycryptodome==3.20.0

# Networking & Async
aiofiles==23.2.1

# Development & Code Quality
black==24.1.1
mypy==1.8.0
pylint==3.0.3
